pyproject.toml
[build-system]
requires = [ "poetry-core>=1.0.0",]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "opsi-cli"
version = "4.3.1.1"
description = "opsi command line interface"
homepage = "https://www.opsi.org"
license = "AGPL-3.0"
authors = [ "uib GmbH <info@uib.de>",]
maintainers = [ "uib GmbH <info@uib.de>",]
[[tool.poetry.packages]]
include = "opsicli"

[[tool.poetry.source]]
name = "uib"
url = "http://pypi.uib.gmbh:8080/simple/"
priority = "primary"

[[tool.poetry.source]]
name = "PyPI"
priority = "primary"

[tool.black]
line-length = 140
target-version = [ "py310",]
use-tabs = true

[tool.isort]
profile = "black"
indent = "\t"

[tool.mypy]
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
follow_imports = "silent"
incremental = false

[tool.ruff]
line-length = 140

[tool.poetry.dependencies]
"ruamel.yaml" = "^0.17"
click = "^8.1"
msgpack = "^1.0"
orjson = "^3.6"
pipreqs = "^0.4"
python = "^3.11,<3.12"
python-opsi-common = ">=4.3,<4.4"
rich-click = "^1.3"
starlette = "^0.28"
websockets = "^11.0"
prompt-toolkit = "^3.0.33"
passlib = "^1.7.4"
pydantic = "^2.5.1"

[tool.poetry.dev-dependencies]
isort = "^5.10"
mypy = "^1.4"
pyinstaller = "^5.5"
pytest = "^7.1"
pytest-cov = "^4.0"
tan = "^23.3"

[tool.poetry.scripts]
opsi-cli = "opsicli.__main__:main"

[tool.ruff.format]
indent-style = "tab"

[tool.poetry.dependencies.pywin32]
platform = "win32"
version = ">=303,<304"

[tool.poetry.group.dev.dependencies]
types-requests = "^2.31.0.10"
ruff = "^0.3.0"
